Separate i18next2js
