+++ lib/index.js
@@ -2,6 +2,7 @@ import po2js from './po2js.js'
 import js2po from './js2po.js'
 import js2i18next from './js2i18next.js'
 import po2i18next from './po2i18next.js'
+import i18next2js from './i18next2js.js'
 import i18next2po from './i18next2po.js'
-export { po2js, js2po, js2i18next, po2i18next, i18next2po }
-export default { po2js, js2po, js2i18next, po2i18next, i18next2po }
+export { po2js, js2po, js2i18next, po2i18next, i18next2js, i18next2po }
+export default { po2js, js2po, js2i18next, po2i18next, i18next2js, i18next2po }
